+++ ocgcore/libduel.cpp
@@ -853,7 +853,7 @@ int32 scriptlib::duel_is_environment(lua_State *L) {
 		pduel->game_field->filter_field_effect(EFFECT_CHANGE_ENVIRONMENT, &eset);
 		if(eset.count) {
 			effect* peffect = eset.get_last();
-			if(code == peffect->get_value() && (playerid == peffect->get_handler_player() || playerid == PLAYER_ALL))
+			if(code == (uint32)peffect->get_value() && (playerid == peffect->get_handler_player() || playerid == PLAYER_ALL))
 				ret = 1;
 		}
 	}
